requirements.txt
black 
autoflake 
isort 
vulture
fastapi[standard]
httpx
google-genai
pydantic_settings
pydantic
Crawl4AI==0.5.0
browser-use
playwright
langchain-google-genai
docker>=6.0.0
redis
pinecone
motor==3.6.0
pymongo
aiohttp